LayoutTests/imported/w3c/web-platform-tests/css/css-values/rex-invalidation-expected.txt

CONSOLE MESSAGE: Error: assert_not_equals: expect the x-height of Ahem and sans-serif to be different got disallowed value 784

FAIL CSS Values and Units Test: rex invalidation Error: assert_not_equals: expect the x-height of Ahem and sans-serif to be different got disallowed value 784
